Authorize in own folder and initial response panel
