+++ app/client/src/features/onlineGoodsAndServices/AuthorizeAPayment/PaymentResponsePanel.tsx
@@ -1,0 +1,3 @@
+export const PaymentResponsePanel = () => {
+  return <div>hello</div>;
+};
