+++ app/client/src/features/onlineGoodsAndServices/AuthorizeAPayment/utils/convertToPaymentResponse.ts
@@ -17,5 +17,6 @@ export function convertToPaymentResponse(
   defaultResponse.paymentMethodType = JSON.parse(values.paymentMethod);
   defaultResponse.requestId = crypto.randomUUID();
   defaultResponse.transactionId = crypto.randomUUID();
+  defaultResponse.transactionDate = new Date().toISOString();
   return defaultResponse;
 }
